// src/domain/interface/config-options.interface.ts
export interface IConfigOptions {
	withCheckFile?: boolean;
	withJavascript?: boolean;
	withJson?: boolean;
	withNest?: boolean;
	withNode?: boolean;
	withPackageJson?: boolean;
	withPerfectionist?: boolean;
	withPrettier?: boolean;
	withReact?: boolean;
	withRegexp?: boolean;
	withSonar?: boolean;
	withStylistic?: boolean;
	withTailwindCss?: boolean;
	withTypeorm?: boolean;
	withTypescript?: boolean;
	withUnicorn?: boolean;
	withYaml?: boolean;
}